requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord.py==1.7.3
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
-e git+https://github.com/forrest-herman/python-packages-common.git@fddb88d2c28abbecea65fe4cc40d2fd58aff9894#egg=google_calendar_integrations&subdirectory=google_calendar_integrations
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
yarl==1.7.2
